Consolidating worker code into src
