+++ package.json
@@ -9,22 +9,24 @@
   "scripts": {
     "axe:ci": "axe --exit --load-delay 1000 `cat axe-urls` --dir ./reports/ --chromedriver-path=\"node_modules/chromedriver/bin/chromedriver\"",
     "build": "./node_modules/.bin/babel src -d ./build/server && ./node_modules/.bin/babel config -d build/config",
-    "deps:local": "yarn install && yarn --cwd frontend install && yarn --cwd worker install",
-    "deps": "yarn install --frozen-lockfile && yarn --cwd frontend install --frozen-lockfile && yarn --cwd worker install --frozen-lockfile",
+    "deps:local": "yarn install && yarn --cwd frontend install",
+    "deps": "yarn install --frozen-lockfile && yarn --cwd frontend install --frozen-lockfile",
     "start:local": "concurrently \"yarn server\" \"yarn client\"",
-    "start:ci": "cross-env POSTGRES_USERNAME=postgres POSTGRES_DB=ttasmarthub TTA_SMART_HUB_URI=http://localhost:3000 concurrently \"yarn start\" \"yarn client\"",
-    "start": "node ./build/server/index.js",
+    "start:ci": "cross-env POSTGRES_USERNAME=postgres POSTGRES_DB=ttasmarthub TTA_SMART_HUB_URI=http://localhost:3000 concurrently \"yarn start:web\" \"yarn client\"",
+    "start:web": "node ./build/server/index.js",
+    "start:worker": "node ./build/server/worker.js",
     "server": "nodemon src/index.js --exec babel-node",
+    "worker": "nodemon src/worker.js --exec babel-node",
     "server:debug": "nodemon src/index.js --exec babel-node --inspect",
     "client": "yarn --cwd frontend start",
     "test": "jest src --runInBand",
     "test:ci": "cross-env JEST_JUNIT_OUTPUT_DIR=reports JEST_JUNIT_OUTPUT_NAME=unit.xml POSTGRES_USERNAME=postgres POSTGRES_DB=ttasmarthub CURRENT_USER_ID=5 CI=true jest src --coverage --reporters=default --reporters=jest-junit",
-    "test:all": "yarn test:ci && yarn --cwd frontend test:ci && yarn --cwd worker test:ci",
+    "test:all": "yarn test:ci && yarn --cwd frontend test:ci",
     "lint": "eslint src",
     "lint:ci": "eslint -f eslint-formatter-multiple src",
-    "lint:all": "yarn lint:ci && yarn --cwd frontend lint:ci && yarn --cwd worker lint:ci",
+    "lint:all": "yarn lint:ci && yarn --cwd frontend lint:ci",
     "lint:fix": "eslint --fix src",
-    "lint:fix:all": "yarn lint:fix && yarn --cwd frontend lint:fix && yarn --cwd workder lint:fix",
+    "lint:fix:all": "yarn lint:fix && yarn --cwd frontend lint:fix",
     "clean": "rm -rf coverage reports frontend/coverage frontend/reports frontend/build",
     "docs:serve": "npx redoc-cli serve -p 5000 docs/openapi/index.yaml",
     "cucumber": "./node_modules/.bin/cucumber-js --publish ./cucumber/features/*.feature -f json:./reports/cucumber_report.json && node ./cucumber/index.js",
@@ -41,17 +43,16 @@
     "db:seed:prod": "node_modules/.bin/sequelize db:seed:all --options-path .production.sequelizerc",
     "db:seed:undo": "node_modules/.bin/sequelize db:seed:undo:all",
     "db:seed:undo:prod": "node_modules/.bin/sequelize db:seed:undo:all --options-path .production.sequelizerc",
-    "docker:deps": "docker-compose run --rm backend yarn install && docker-compose run --rm frontend yarn install && docker-compose run --rm worker yarn install",
+    "docker:deps": "docker-compose run --rm backend yarn install && docker-compose run --rm frontend yarn install",
     "docker:reset": "./bin/reset-all",
     "docker:start": "docker-compose up",
     "docker:stop": "docker-compose down",
     "docker:test": "./bin/run-tests",
     "docker:test:be": "docker-compose run --rm backend yarn test",
-    "docker:lint": "docker-compose run --rm backend yarn lint:ci && docker-compose run --rm frontend yarn lint:ci && docker-compose run --rm worker lint:ci",
-    "docker:lint:fix": "docker-compose run --rm backend yarn lint:fix && docker-compose run --rm frontend yarn lint:fix && docker-compose run --rm worker lint:fix",
+    "docker:lint": "docker-compose run --rm backend yarn lint:ci && docker-compose run --rm frontend yarn lint:ci",
+    "docker:lint:fix": "docker-compose run --rm backend yarn lint:fix && docker-compose run --rm frontend yarn lint:fix",
     "docker:shell:frontend": "docker-compose run --rm frontend /bin/bash",
     "docker:shell:backend": "docker-compose run --rm backend /bin/bash",
-    "docker:shell:worker": "docker-compose run --rm worker /bin/bash",
     "docker:db:migrate": "docker-compose run --rm backend node_modules/.bin/sequelize db:migrate",
     "docker:db:migrate:undo": "docker-compose run --rm backend node_modules/.bin/sequelize db:migrate:undo",
     "docker:db:seed": "docker-compose run --rm backend yarn db:seed",
@@ -90,8 +91,7 @@
   },
   "jest": {
     "testPathIgnorePatterns": [
-      "<rootDir>/frontend/",
-      "<rootDir>/worker/"
+      "<rootDir>/frontend/"
     ],
     "coverageThreshold": {
       "global": {
@@ -159,6 +159,7 @@
     "express-unless": "^0.5.0",
     "express-winston": "^4.0.5",
     "file-type": "^16.1.0",
+    "form-data": "^4.0.0",
     "helmet": "^4.1.0",
     "http-codes": "^1.0.0",
     "lodash": "^4.17.20",
@@ -171,6 +172,7 @@
     "puppeteer-select": "^1.0.3",
     "sequelize": "^5.22.3",
     "sequelize-cli": "^6.2.0",
+    "throng": "^5.0.0",
     "url-join": "^4.0.1",
     "uuid": "^8.3.2",
     "winston": "^3.3.3",
