+++ src/app.js
@@ -13,7 +13,7 @@ import updateGrantsGrantees from './lib/updateGrantsGrantees';
 
 import findOrCreateUser from './services/accessValidation';
 
-import logger, { auditLogger, requestLogger } from './logger';
+import { logger, auditLogger, requestLogger } from './logger';
 
 const app = express();
