--- worker/config/config.js
@@ -1,27 +1,0 @@
-require('dotenv').config();
-
-module.exports = {
-  development: {
-    username: process.env.POSTGRES_USERNAME,
-    password: process.env.POSTGRES_PASSWORD,
-    database: process.env.POSTGRES_DB,
-    host: process.env.POSTGRES_HOST,
-    dialect: 'postgres',
-  },
-  test: {
-    username: process.env.POSTGRES_USERNAME,
-    password: process.env.POSTGRES_PASSWORD,
-    database: process.env.POSTGRES_DB,
-    host: process.env.POSTGRES_HOST,
-    dialect: 'postgres',
-    logging: false,
-  },
-  production: {
-    use_env_variable: 'DATABASE_URL',
-    username: process.env.POSTGRES_USERNAME,
-    password: process.env.POSTGRES_PASSWORD,
-    database: process.env.POSTGRES_DB,
-    host: process.env.POSTGRES_HOST,
-    dialect: 'postgres',
-  },
-};
